feat: #36 admin-main 페이지 스타일링
