+++ src/App.vue
@@ -6,6 +6,13 @@
 
 <style lang="scss">
 @import '@/assets/scss/_index.scss';
+
+#app {
+  font-family: Avenir, Helvetica, Arial, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
 </style>
 
 <script lang="ts">
