--- src/assets/scss/layouts/admin-header.scss
@@ -1,15 +1,0 @@
-.admin-header {
-  //width: 100%;
-  background-color: white;
-
-  &__title {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    padding: 30px 30px 0 30px;
-  }
-
-  &__breadcrumb {
-    padding: 30px;
-  }
-}
