+++ src/assets/scss/layouts/empty-layout.scss
@@ -1,7 +1,7 @@
 .empty-layout {
+  width: 90vw;
   display: flex;
   flex-direction: column;
-  width: 90vw;
   align-items: center;
   margin: auto;
 }
