+++ src/assets/scss/pages/page-admin-main.scss
@@ -1,0 +1,53 @@
+
+.admin-main__body {
+  width: 100%;
+}
+
+.button {
+  margin: 20px 0 20px;
+}
+
+.button--create {
+  a {
+    color: #5f6368;
+  }
+}
+
+.admin-main__body {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.table {
+  width: 100%;
+  border-collapse: collapse;
+  color: #5f6368;
+
+  a {
+    color: #5f6368;
+  }
+
+  &__thead {
+    height: 40px;
+    text-align: center;
+    background-color: dodgerblue;
+    color: #fff;
+  }
+
+  &__tbody {
+    text-align: center;
+
+    tr:last-child {
+      border: none;
+    }
+  }
+
+  &__tr {
+    height: 30px;
+    border-bottom: 1px solid #eee;
+  }
+
+
+}
